Move computation of Blosc2 array args to HDF5 support module
